+++ spec/app/models/ordered_post.rb
@@ -1,0 +1,6 @@
+class OrderedPost
+  include Mongoid::Document
+  field :title, :type => String
+  field :rating, :type => Integer
+  belongs_to :person
+end